add DNS cache
